Clarify soversion
The soversion of libkdcraw is currently intended to be set to "30.0.0". This is quite an unusual value, it seems to be so for historical reasons (from when the library was part of digikam). In practice, this didn't work due to a typo, so the soversion was in fact 5 (the major lib version).
Since this is no longer released nor used by digikam (which nowadays ships its own libraw wrapping code), set the soversion directly to 5 (so it can be bumped when there are API breaking changes)

Differential Revision: https://phabricator.kde.org/D15149